Update to ember-cli v1.13.1
