+++ .npmignore
@@ -10,5 +10,6 @@ dist/
 .npmignore
 **/.gitkeep
 bower.json
+ember-cli-build.js
 Brocfile.js
 testem.json
