+++ bower.json
@@ -1,22 +1,22 @@
 {
   "name": "ember-jsonapi-resources",
   "dependencies": {
-    "ember": "~1.13.2",
+    "ember": "~1.13.3",
     "ember-cli-shims": "ember-cli/ember-cli-shims#0.0.3",
     "ember-cli-test-loader": "ember-cli-test-loader#0.1.3",
-    "ember-load-initializers": "ember-cli/ember-load-initializers#0.1.4",
-    "ember-qunit": "0.3.3",
+    "ember-load-initializers": "ember-cli/ember-load-initializers#0.1.5",
+    "ember-qunit": "0.4.1",
     "ember-qunit-notifications": "0.0.7",
-    "ember-resolver": "~0.1.15",
+    "ember-resolver": "~0.1.18",
     "jquery": "^1.11.1",
     "loader.js": "ember-cli/loader.js#3.2.0",
-    "qunit": "~1.17.1",
     "fetch": "~0.9.0",
     "es6-promise": "~2.2.0",
     "blanket": "~1.1.5"
   },
   "devDependencies": {
     "sinon": "http://sinonjs.org/releases/sinon-1.15.0.js",
-    "es5-shim": "^4.0.5"
+    "es5-shim": "^4.0.5",
+    "qunit": "~1.17.1"
   }
 }
