+++ coverage.json
@@ -5,8 +5,8 @@
     "passes": 115,
     "pending": 0,
     "failures": 0,
-    "start": "2015-06-22T06:41:46.493Z",
-    "end": "2015-06-22T06:41:50.473Z"
+    "start": "2015-07-10T19:03:02.184Z",
+    "end": "2015-07-10T19:03:06.167Z"
   },
   "coverage": {
     "total": {
