+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ember-jsonapi-resources",
-  "version": "0.4.2",
+  "version": "0.4.3",
   "description": "Lightweight persistence for an Ember CLI app using the JSON API 1.0 specification.",
   "directories": {
     "doc": "doc",
@@ -23,15 +23,17 @@
     "broccoli-filter": "^0.1.14",
     "broccoli-static-compiler": "^0.2.1",
     "ember-buffered-proxy": "0.5.0",
-    "ember-cli": "^0.2.7",
-    "ember-cli-app-version": "0.3.3",
+    "ember-cli": "^1.13.1",
+    "ember-cli-app-version": "0.4.0",
     "ember-cli-blanket": "0.5.2",
     "ember-cli-content-security-policy": "0.4.0",
     "ember-cli-dependency-checker": "^1.0.0",
-    "ember-cli-htmlbars": "0.7.6",
+    "ember-cli-htmlbars": "0.7.9",
+    "ember-cli-htmlbars-inline-precompile": "^0.1.1",
     "ember-cli-ic-ajax": "0.1.1",
     "ember-cli-inject-live-reload": "^1.3.0",
-    "ember-cli-qunit": "0.3.13",
+    "ember-cli-qunit": "0.3.15",
+    "ember-cli-release": "0.2.3",
     "ember-cli-uglify": "^1.0.1",
     "ember-disable-prototype-extensions": "^1.0.0",
     "ember-disable-proxy-controllers": "^1.0.0",
